Move is_niceness logic out of clap utils to reduce dependencies..
..for solana-keygen
